Use Uri as input for FormLoaderTask
