+++ collect_app/src/androidTest/java/org/odk/collect/android/instrumented/tasks/FormLoaderTaskTest.java
@@ -2,6 +2,7 @@ package org.odk.collect.android.instrumented.tasks;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.notNullValue;
+import static org.junit.Assert.fail;
 import static org.mockito.Mockito.mock;
 
 import android.app.Application;
@@ -66,54 +67,58 @@ public class FormLoaderTaskTest {
     // Validate the use of CSV files as secondary instances accessed through "jr://file-csv"
     @Test
     public void loadFormWithSecondaryCSV() throws Exception {
-        final String formPath = storagePathProvider.getOdkDirPath(StorageSubdirectory.FORMS) + File.separator + SECONDARY_INSTANCE_EXTERNAL_CSV_FORM;
-        FormLoaderTask formLoaderTask = new FormLoaderTask(formPath, formPath, null, null, formEntryControllerFactory, mock());
-        FormLoaderTask.FECWrapper wrapper = formLoaderTask.executeSynchronously();
-        Assert.assertNotNull(wrapper);
+        fail();
+//        final String formPath = storagePathProvider.getOdkDirPath(StorageSubdirectory.FORMS) + File.separator + SECONDARY_INSTANCE_EXTERNAL_CSV_FORM;
+//        FormLoaderTask formLoaderTask = new FormLoaderTask(formPath, formPath, null, null, formEntryControllerFactory, mock());
+//        FormLoaderTask.FECWrapper wrapper = formLoaderTask.executeSynchronously();
+//        Assert.assertNotNull(wrapper);
     }
 
     // Validate the use of a CSV file externally accessed through search/pulldata
     @Test
     public void loadSearchFromExternalCSV() throws Exception {
-        final String formPath = storagePathProvider.getOdkDirPath(StorageSubdirectory.FORMS) + File.separator + SIMPLE_SEARCH_EXTERNAL_CSV_FORM;
-        FormLoaderTask formLoaderTask = new FormLoaderTask(formPath, formPath, null, null, formEntryControllerFactory, mock());
-        FormLoaderTask.FECWrapper wrapper = formLoaderTask.executeSynchronously();
-        assertThat(wrapper, notNullValue());
+        fail();
+//        final String formPath = storagePathProvider.getOdkDirPath(StorageSubdirectory.FORMS) + File.separator + SIMPLE_SEARCH_EXTERNAL_CSV_FORM;
+//        FormLoaderTask formLoaderTask = new FormLoaderTask(formPath, formPath, null, null, formEntryControllerFactory, mock());
+//        FormLoaderTask.FECWrapper wrapper = formLoaderTask.executeSynchronously();
+//        assertThat(wrapper, notNullValue());
     }
 
     @Test
     public void loadSearchFromexternalCsvLeavesFileUnchanged() throws Exception {
-        final String formPath = storagePathProvider.getOdkDirPath(StorageSubdirectory.FORMS) + File.separator + SIMPLE_SEARCH_EXTERNAL_CSV_FORM;
-        FormLoaderTask formLoaderTask = new FormLoaderTask(formPath, formPath, null, null, formEntryControllerFactory, mock());
-        FormLoaderTask.FECWrapper wrapper = formLoaderTask.executeSynchronously();
-        Assert.assertNotNull(wrapper);
-        Assert.assertNotNull(wrapper.getController());
-
-        File mediaFolder = wrapper.getController().getMediaFolder();
-        File importedCSV = new File(mediaFolder + File.separator + SIMPLE_SEARCH_EXTERNAL_CSV_FILE);
-        Assert.assertTrue("Expected the imported CSV file to remain unchanged", importedCSV.exists());
+        fail();
+//        final String formPath = storagePathProvider.getOdkDirPath(StorageSubdirectory.FORMS) + File.separator + SIMPLE_SEARCH_EXTERNAL_CSV_FORM;
+//        FormLoaderTask formLoaderTask = new FormLoaderTask(formPath, formPath, null, null, formEntryControllerFactory, mock());
+//        FormLoaderTask.FECWrapper wrapper = formLoaderTask.executeSynchronously();
+//        Assert.assertNotNull(wrapper);
+//        Assert.assertNotNull(wrapper.getController());
+//
+//        File mediaFolder = wrapper.getController().getMediaFolder();
+//        File importedCSV = new File(mediaFolder + File.separator + SIMPLE_SEARCH_EXTERNAL_CSV_FILE);
+//        Assert.assertTrue("Expected the imported CSV file to remain unchanged", importedCSV.exists());
     }
 
     // Validate that importing external data multiple times does not fail due to side effects from import
     @Test
     public void loadSearchFromExternalCSVmultipleTimes() throws Exception {
-        final String formPath = storagePathProvider.getOdkDirPath(StorageSubdirectory.FORMS) + File.separator + SIMPLE_SEARCH_EXTERNAL_CSV_FORM;
-        // initial load with side effects
-        FormLoaderTask formLoaderTask = new FormLoaderTask(formPath, formPath, null, null, formEntryControllerFactory, mock());
-        FormLoaderTask.FECWrapper wrapper = formLoaderTask.executeSynchronously();
-        Assert.assertNotNull(wrapper);
-        Assert.assertNotNull(wrapper.getController());
-
-        File mediaFolder = wrapper.getController().getMediaFolder();
-        File dbFile = new File(mediaFolder + File.separator + SIMPLE_SEARCH_EXTERNAL_DB_FILE);
-        Assert.assertTrue(dbFile.exists());
-        long dbLastModified = dbFile.lastModified();
-
-        // subsequent load should succeed despite side effects from import
-        formLoaderTask = new FormLoaderTask(formPath, formPath, null, null, formEntryControllerFactory, mock());
-        wrapper = formLoaderTask.executeSynchronously();
-        Assert.assertNotNull(wrapper);
-        Assert.assertNotNull(wrapper.getController());
-        Assert.assertEquals("expected file modification timestamp to be unchanged", dbLastModified, dbFile.lastModified());
+        fail();
+//        final String formPath = storagePathProvider.getOdkDirPath(StorageSubdirectory.FORMS) + File.separator + SIMPLE_SEARCH_EXTERNAL_CSV_FORM;
+//        // initial load with side effects
+//        FormLoaderTask formLoaderTask = new FormLoaderTask(formPath, formPath, null, null, formEntryControllerFactory, mock());
+//        FormLoaderTask.FECWrapper wrapper = formLoaderTask.executeSynchronously();
+//        Assert.assertNotNull(wrapper);
+//        Assert.assertNotNull(wrapper.getController());
+//
+//        File mediaFolder = wrapper.getController().getMediaFolder();
+//        File dbFile = new File(mediaFolder + File.separator + SIMPLE_SEARCH_EXTERNAL_DB_FILE);
+//        Assert.assertTrue(dbFile.exists());
+//        long dbLastModified = dbFile.lastModified();
+//
+//        // subsequent load should succeed despite side effects from import
+//        formLoaderTask = new FormLoaderTask(formPath, formPath, null, null, formEntryControllerFactory, mock());
+//        wrapper = formLoaderTask.executeSynchronously();
+//        Assert.assertNotNull(wrapper);
+//        Assert.assertNotNull(wrapper.getController());
+//        Assert.assertEquals("expected file modification timestamp to be unchanged", dbLastModified, dbFile.lastModified());
     }
 }
